+++ src/components/Text/Text.stories.tsx
@@ -4,7 +4,7 @@ import React from 'react'
 import { Text } from './Text'
 
 export default {
-  title: 'Text',
+  title: 'Text（テキスト）/Text',
   component: Text,
 }